changes for interpolation. Still not in a working state, but everything compiles. Just need to implement Rotate and Translate, then should be good to go. Misc other changes to get the MouseEvents to compile. Added mingw makefiles.

+++ trunk/src/interpolation/TimedInterpolation.cpp
@@ -10,7 +10,7 @@ TimedInterpolation::TimedInterpolation(int milliseconds, InterpolationListener *
 {
     _time = milliseconds;
 
-    _iterationsRemaining = _time/(double)Timer::GetInstance()->getInterval();
+    _iterationsRemaining = (int)(_time/(double)Timer::GetInstance()->getInterval());
 
     addListener(listener);
 }
@@ -43,6 +43,4 @@ void TimedInterpolation::start()
     Timer::GetInstance()->addListener(this);
 }
 
-
 }
-
